requirements.txt
# Core ML
ultralytics

# Computer vision utilities
supervision
albumentations
Pillow
numpy

# API
fastapi
python-multipart
uvicorn

# Configuration
<<<<<<< Updated upstream
pydantic
pydantic-settings
=======
pydantic==2.12.5
pydantic-settings==2.13.
>>>>>>> Stashed changes
